Changes VideoTexture to StreamingTexture, to better reflect what it actually does.
